make pr lighter
