+++ Sources/Common/Communication/EventStorage.swift
@@ -79,8 +79,8 @@ actor EventStorageManager: EventStorage {
                 logger.error("Warning: Could not decode event at \(url). Error: \(error)")
             }
         }
-        // Sort events based on their storageId (which includes a timestamp).
-        return events.sorted(by: { $0.storageId < $1.storageId })
+        // Sort events based on their timestamp.
+        return events.sorted(by: { $0.timestamp < $1.timestamp })
     }
 
     // Removes a specific event file
